try to consolidate vitest setup
