--- apps/spruce/config/vitest/setupTests.ts
@@ -1,72 +1,0 @@
-import "@testing-library/jest-dom/vitest";
-import "vitest-canvas-mock";
-import { vi } from "vitest";
-
-// @ts-expect-error: Returning a basic string is acceptable for tests.
-window.crypto.randomUUID = (() => {
-  let value = 0;
-  return () => {
-    value += 1;
-    return value.toString();
-  };
-})();
-
-// @ts-expect-error: Workaround for a bug in @testing-library/react.
-// It prevents Vitest's fake timers from functioning with user-event.
-// https://github.com/testing-library/react-testing-library/issues/1197
-globalThis.jest = vi;
-
-beforeEach(() => {
-  // LeafyGreen tables require an IntersectionObserver.
-  const mockIntersectionObserver = vi.fn(function () {
-    return {
-      observe: vi.fn(),
-      unobserve: vi.fn(),
-      disconnect: vi.fn(),
-    };
-  });
-  vi.stubGlobal("IntersectionObserver", mockIntersectionObserver);
-
-  const mockResizeObserver = vi.fn(function () {
-    return {
-      observe: vi.fn(),
-      unobserve: vi.fn(),
-      disconnect: vi.fn(),
-    };
-  });
-  vi.stubGlobal("ResizeObserver", mockResizeObserver);
-
-  // LeafyGreen modals require stubbing dialog
-  // https://github.com/mongodb/leafygreen-ui/blob/22b4d8200b132f24b4cd1a7e4d99b0372ed6fd58/packages/modal/src/utils/getTestUtils.spec.tsx#L36-L54
-  HTMLDialogElement.prototype.show = vi.fn(function mock(
-    this: HTMLDialogElement,
-  ) {
-    this.open = true;
-  });
-
-  HTMLDialogElement.prototype.showModal = vi.fn(function mock(
-    this: HTMLDialogElement,
-  ) {
-    this.open = true;
-  });
-
-  HTMLDialogElement.prototype.close = vi.fn(function mock(
-    this: HTMLDialogElement,
-  ) {
-    this.open = false;
-  });
-
-  // jsdom 28 recognizes the popover attribute (hiding elements by default)
-  // but doesn't fully implement showPopover/hidePopover.
-  HTMLElement.prototype.showPopover = vi.fn(function mock(this: HTMLElement) {
-    this.style.display = "block";
-  });
-
-  HTMLElement.prototype.hidePopover = vi.fn(function mock(this: HTMLElement) {
-    this.style.display = "";
-  });
-});
-
-afterEach(() => {
-  vi.unstubAllGlobals();
-});
