+++ packages/fungi/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../../tsconfig.base.json",
   "include": [
+    "../lib/vitest.d.ts",
     "src",
     "test_utils/index.tsx",
     "test_utils/custom-queries.ts",
